requirements.txt
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
